feat(command-palette): implement open URL and search web commands with validation
