+++ browser-features/chrome/common/command-palette/multiInputCommand/index.ts
@@ -1,0 +1,4 @@
+// SPDX-License-Identifier: MPL-2.0
+
+export { openUrlCommand } from "./open-url.ts";
+export { searchWebCommand } from "./search-web.ts";
